backend/requirements.txt
arxiv
python-dotenv
openai
Flask
Flask-Cors
python-dateutil